% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{findBigAvalues}
\alias{findBigAvalues}
\title{TTTSizer}
\usage{
findBigAvalues(p0Index, hIndex, phiVector, kernelCube, poliCube)
}
\arguments{
\item{p0Index}{- An index number refering to the p0 value}

\item{hIndex}{- An index number refering to the bandwith}

\item{phiVector}{- The vector with the phi values}

\item{kernelCube}{- The cube with all the Kh(pi-p0) values}

\item{poliCube}{- The cube with all the (pi-p0)^n values}
}
\value{
Return a vector with the (A0,A1,A2, A3) values
}
\description{
For a given set of data, find the A matrix of dimession i x 1
}
\note{
The function needs you to calculate first the following constants

    \enumerate{
        \item - a kernel cube with all the K_h(pi-p0) values
        \item - a polinomial cube with all the (pi-p0)^n values
        \item - a vector with all the phi(xis) values
    }
}
\seealso{
findThetas
}
