% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{findThetasCubic}
\alias{findThetasCubic}
\title{TTTSizer}
\usage{
findThetasCubic(bigAs, littleAs)
}
\arguments{
\item{bigAs}{vector with the big As associated with that p0,x,p,h and kernel}

\item{littleAs}{vector with the little As associated with that p0, p,h and kernel}
}
\value{
c(NaN, NaN, NaN) - If is not a Cramer system
        c(theta0, theta1, theta2)    - Otherwise
}
\description{
Find thetas for a cubic method.
    For a given set of data, find the variable matrix solution of theta0, theta1, and theta2
}
