% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generatePolinomialCube}
\alias{generatePolinomialCube}
\title{TTTSizer}
\usage{
generatePolinomialCube(piVector, p0Vector)
}
\arguments{
\item{piVector}{The vector with all the frecuencies (can be a vector of size one)}

\item{p0Vector}{The vector with only the p0 values (can be a vector of size one)}
}
\value{
A list of matrix with all the(pi-p0)^n combinations pre-calculated.
    The index of the list is n, the row of the matrix is pi and the column is p0
}
\description{
For a given vector of pi's and generate all possible (pi-p0)^n combinations.
    Where n=0,1,2,3,4,5,6
}
