% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{kernelFunction}
\alias{kernelFunction}
\title{TTTSizer}
\usage{
kernelFunction(x, kernel = "gaussian")
}
\arguments{
\item{x}{(float) A value for the symetric kernel function}

\item{kernel}{The selected kernel function

              \describe{
                  \item{gaussian}{exp(-(x^2) / 2) / sqrt(2 * pi)}
                  \item{biweight}{(abs(x)<=1)*15/16*(1-x^2)^2}
                  \item{triweight}{(abs(x)<=1)*35/32*(1-x^2)^3}
                  \item{epanechnikov}{(abs(x)<=1)*((1-x^2)*3/4)}
               }}
}
\value{
(float) The kernel function result.
}
\description{
For a given number, return the value of the kernel function, with the specified kernel
}
\examples{
kernelFunction(1.5)
          kernelFunction(0, kernel="biweight")

}
