% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{summaryColors}
\alias{summaryColors}
\title{TTTSizer}
\usage{
summaryColors(sizerData)
}
\arguments{
\item{sizerData}{The summary returned by the TTT function}
}
\value{
A 12 x 2 dataframe with the color info summarized.
}
\description{
After running a TTT function, summarize the results of each SiZer map into a dataframe.
             - The first column of the dataframe is the color name
             - The second column of the dataframe is proportion of pixels for that SiZer map.
             - Notice that rows are group 4 by 4. The first 4 is the SiZer-0, the next 4 SiZer-1, and the last 4 the SiZer-2
}
